fix: fix test cases
